Move client resources to dedicated pkg
