+++ internal/auth0/client/data_source_test.go
@@ -1,4 +1,4 @@
-package provider
+package client_test
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/template"
 )
@@ -35,7 +36,7 @@ func TestAccDataClientByName(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories:         ProviderTestFactories(httpRecorder),
+		ProviderFactories:         provider.TestFactories(httpRecorder),
 		PreventPostDestroyRefresh: true,
 		Steps: []resource.TestStep{
 			{
@@ -56,7 +57,7 @@ func TestAccDataClientById(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories:         ProviderTestFactories(httpRecorder),
+		ProviderFactories:         provider.TestFactories(httpRecorder),
 		PreventPostDestroyRefresh: true,
 		Steps: []resource.TestStep{
 			{
